+++ CHANGELOG.md
@@ -24,3 +24,4 @@
 # 0.2.0
 * Issue: Change minimum time interval of appointment from 30 minutes to 15 minutes
 * Issue: Add tap gesture to enable and disable edition mode
+* Issue: Make possible to build child widget of Appointment based on duration of appointment